Add tokenizer

+++ parser/types.go
@@ -2,13 +2,13 @@ package parser
 
 const (
 	//PTypeNone NA
-	PTypeNone int = -1
+	PTypeNone int8 = -1
 	//PTypeAddition Addition.
-	PTypeAddition int = 0
+	PTypeAddition int8 = 0
 	//PTypeMultiplication Multiplication.
-	PTypeMultiplication int = 1
+	PTypeMultiplication int8 = 1
 	//PTypeDivision Division.
-	PTypeDivision int = 2
+	PTypeDivision int8 = 2
 	//PTypeSubtraction Subtraction.
-	PTypeSubtraction int = 3
+	PTypeSubtraction int8 = 3
 )
